Implement `OpenFont` method in `XBufferProto` and enhance response handling for `QueryFont` and `ListHosts`. Fix struct definitions for `OpenFontType` and `QueryFontResponse`.

+++ Src/Models/Requests/OpenFontType.cs
@@ -6,12 +6,12 @@ namespace Xcsb.Models.Requests;
 
 [StructLayout(LayoutKind.Sequential, Pack = 1, Size = 12)]
 [method: MethodImpl(MethodImplOptions.AggressiveInlining)]
-internal readonly struct OpenFontType(uint fontId, ushort fontLength)
+internal readonly struct OpenFontType(uint fontId, int fontLength)
 {
     public readonly Opcode opcode = Opcode.OpenFont;
     private readonly byte _pad0;
     public readonly ushort Length = (ushort)(3 + fontLength.AddPadding() / 4);
     public readonly uint FontId = fontId;
-    public readonly ushort FontLength = fontLength;
+    public readonly ushort FontLength = (ushort)fontLength;
     private readonly ushort _pad1;
 }
